backend/requirements.txt
fastapi
uvicorn[standard]
PyPDF2
python-dotenv
langdetect
google-generativeai>=0.3.0
reportlab
python-multipart
google-genai